@@ -13,6 +13,10 @@ gmail_oauth2_require = [
     'python-social-auth',
 ]
 
+office365_oauth2_require = [
+    'O365',
+]
+
 setup(
     name='django-mailbox',
     version=version_string,
@@ -25,29 +29,28 @@ setup(
     author='Adam Coddington',
     author_email='me@adamcoddington.net',
     extras_require={
-        'gmail-oauth2': gmail_oauth2_require
+        'gmail-oauth2': gmail_oauth2_require,
+        'office365-oauth2': office365_oauth2_require
     },
-    python_requires=">=2.7",
+    python_requires=">=3.8",
     classifiers=[
         'Framework :: Django',
         'Intended Audience :: Developers',
         'License :: OSI Approved :: MIT License',
         'Operating System :: OS Independent',
         'Framework :: Django',
-        'Framework :: Django :: 1.11',
-        'Framework :: Django :: 2.0',
-        'Framework :: Django :: 2.1',
-        'Framework :: Django :: 2.2',
-        'Framework :: Django :: 3.0',
+        'Framework :: Django :: 3.2',
+        'Framework :: Django :: 4.0',
+        'Framework :: Django :: 4.1',
+        'Framework :: Django :: 4.2',
+        'Framework :: Django :: 5.0',
         'Programming Language :: Python',
-        'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.4',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
         'Topic :: Communications :: Email',
         'Topic :: Communications :: Email :: Post-Office',
         'Topic :: Communications :: Email :: Post-Office :: IMAP',
@@ -56,7 +59,4 @@ setup(
     ],
     packages=find_packages(),
     include_package_data=True,
-    install_requires=[
-        'six>=1.6.1'
-    ]
 )
