@@ -1,3 +1,3 @@
-__version__ = '4.8.2'
+__version__ = '4.10.1'
 
 default_app_config = 'django_mailbox.apps.MailBoxConfig'
