@@ -15,7 +15,6 @@ import email
 import logging
 import mimetypes
 import os.path
-import sys
 import uuid
 from tempfile import NamedTemporaryFile
 
@@ -31,7 +30,7 @@ from django_mailbox import utils
 from django_mailbox.signals import message_received
 from django_mailbox.transports import Pop3Transport, ImapTransport, \
     MaildirTransport, MboxTransport, BabylTransport, MHTransport, \
-    MMDFTransport, GmailImapTransport
+    MMDFTransport, GmailImapTransport, Office365Transport
 
 logger = logging.getLogger(__name__)
 
@@ -42,7 +41,7 @@ class MailboxQuerySet(models.QuerySet):
         for mailbox in self.all():
             logger.debug('Receiving mail for %s' % mailbox)
             count += sum(1 for i in mailbox.get_new_mail())
-        logger.debug('Received %d %s.', count, 'mails' if count != 1 else 'mail')
+        logger.debug('Received %d %s.', count, 'mail(s)')
 
 
 class MailboxManager(models.Manager):
@@ -131,6 +130,13 @@ class Mailbox(models.Model):
     objects = MailboxManager()
     active_mailboxes = ActiveMailboxManager()
 
+    class Meta:
+        verbose_name = _('Mailbox')
+        verbose_name_plural = _('Mailboxes')
+
+    def __str__(self):
+        return self.name
+
     @property
     def _protocol_info(self):
         return urlparse(self.uri)
@@ -197,6 +203,30 @@ class Mailbox(models.Model):
             return None
         return folder[0]
 
+    @property
+    def client_id(self):
+        """Returns (if specified) the client id for Office365."""
+        client_id = self._query_string.get('client_id', None)
+        if not client_id:
+            return None
+        return client_id[0]
+
+    @property
+    def client_secret(self):
+        """Returns (if specified) the client secret for Office365."""
+        client_secret = self._query_string.get('client_secret', None)
+        if not client_secret:
+            return None
+        return client_secret[0]
+
+    @property
+    def tenant_id(self):
+        """Returns (if specified) the tenant id for Office365."""
+        tenant_id = self._query_string.get('tenant_id', None)
+        if not tenant_id:
+            return None
+        return tenant_id[0]
+
     def get_connection(self):
         """Returns the transport instance for this mailbox.
 
@@ -231,6 +261,14 @@ class Mailbox(models.Model):
                 ssl=self.use_ssl
             )
             conn.connect(self.username, self.password)
+        elif self.type == 'office365':
+            conn = Office365Transport(
+                self.location,
+                self.username,
+                folder=self.folder,
+                archive=self.archive
+            )
+            conn.connect(self.client_id, self.client_secret, self.tenant_id)
         elif self.type == 'maildir':
             conn = MaildirTransport(self.location)
         elif self.type == 'mbox':
@@ -268,7 +306,10 @@ class Mailbox(models.Model):
         settings = utils.get_settings()
 
         new = EmailMessage()
-        if msg.is_multipart():
+        if (
+            msg.is_multipart()
+            and 'attachment' not in msg.get('Content-Disposition', '')
+        ):
             for header, value in msg.items():
                 new[header] = value
             for part in msg.get_payload():
@@ -309,11 +350,21 @@ class Mailbox(models.Model):
 
             attachment = MessageAttachment()
 
+            if msg.get_content_type() == 'message/rfc822':
+                attachment_payloads = msg.get_payload()
+                if len(attachment_payloads) != 1:
+                    raise AssertionError(
+                        "Attachment of type 'message/rfc822' "
+                        'must have exactly 1 payload.'
+                    )
+                attachment_payload = attachment_payloads[0].as_bytes()
+            else:
+                attachment_payload = msg.get_payload(decode=True)
             attachment.document.save(
                 uuid.uuid4().hex + extension,
                 ContentFile(
                     BytesIO(
-                        msg.get_payload(decode=True)
+                        attachment_payload
                     ).getvalue()
                 )
             )
@@ -424,26 +475,41 @@ class Mailbox(models.Model):
 
     def get_new_mail(self, condition=None):
         """Connect to this transport and fetch new messages."""
-        new_mail = []
         connection = self.get_connection()
         if not connection:
             return
-        for message in connection.get_message(condition):
-            msg = self.process_incoming_message(message)
-            if not msg is None:
-                yield msg
-        self.last_polling = now()
-        if django.VERSION >= (1, 5):  # Django 1.5 introduces update_fields
-            self.save(update_fields=['last_polling'])
-        else:
-            self.save()
+        try:
+            for message in connection.get_message(condition):
+                msg = self.process_incoming_message(message)
+                if msg is not None:
+                    yield msg
+            self.last_polling = now()
+            if django.VERSION >= (1, 5):  # Django 1.5 introduces update_fields
+                self.save(update_fields=['last_polling'])
+            else:
+                self.save()
+        finally:
+            connection.close()
 
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = _('Mailbox')
-        verbose_name_plural = _('Mailboxes')
+    @staticmethod
+    def get_new_mail_all_mailboxes(args=None):
+        mailboxes = Mailbox.active_mailboxes.all()
+        if args:
+            mailboxes = mailboxes.filter(
+                name=' '.join(args)
+            )
+        for mailbox in mailboxes:
+            logger.info(
+                'Gathering messages for %s',
+                mailbox.name
+            )
+            messages = mailbox.get_new_mail()
+            for message in messages:
+                logger.info(
+                    'Received %s (from %s)',
+                    message.subject,
+                    message.from_address
+                )
 
 
 class IncomingMessageManager(models.Manager):
@@ -534,6 +600,7 @@ class Message(models.Model):
     eml = models.FileField(
         _('Raw message contents'),
         null=True,
+        blank=True,
         upload_to="messages",
         help_text=_('Original full content of message')
     )
@@ -588,14 +655,24 @@ class Message(models.Model):
         return addresses
 
     def reply(self, message):
-        """Sends a message as a reply to this message instance.
+        """Sends an EmailMessage as a reply to this message instance::
+
+            from django.core.mail import EmailMessage
+
+            message.reply(
+                EmailMessage(subject="pong", body="pongpong")
+            )
 
         Although Django's e-mail processing will set both Message-ID
         and Date upon generating the e-mail message, we will not be able
         to retrieve that information through normal channels, so we must
         pre-set it.
-
         """
+        from django.core.mail import EmailMessage as DjangoEmailMessage
+
+        if not isinstance(message, DjangoEmailMessage):
+            raise ValueError('Message must be an instance of django.core.mail.EmailMessage')
+
         if not message.from_email:
             if self.mailbox.from_email:
                 message.from_email = self.mailbox.from_email
@@ -632,7 +709,6 @@ class Message(models.Model):
     def _rehydrate(self, msg):
         new = EmailMessage()
         settings = utils.get_settings()
-
         if msg.is_multipart():
             for header, value in msg.items():
                 new[header] = value
@@ -651,24 +727,24 @@ class Message(models.Model):
                 if encoding and encoding.lower() == 'quoted-printable':
                     # Cannot use `email.encoders.encode_quopri due to
                     # bug 14360: http://bugs.python.org/issue14360
-                    output = BytesIO()
-                    encode_quopri(
-                        BytesIO(
-                            attachment.document.read()
-                        ),
-                        output,
-                        quotetabs=True,
-                        header=False,
-                    )
-                    new.set_payload(
-                        output.getvalue().decode().replace(' ', '=20')
-                    )
+                    with open(attachment.document.path, 'rb') as f:
+                        output = BytesIO()
+                        encode_quopri(
+                            BytesIO(
+                                f.read()
+                            ),
+                            output,
+                            quotetabs=True,
+                            header=False,
+                        )
+                        new.set_payload(
+                            output.getvalue().decode().replace(' ', '=20')
+                        )
                     del new['Content-Transfer-Encoding']
                     new['Content-Transfer-Encoding'] = 'quoted-printable'
                 else:
-                    new.set_payload(
-                        attachment.document.read()
-                    )
+                    with open(attachment.document.path, 'rb') as f:
+                        new.set_payload(f.read())
                     del new['Content-Transfer-Encoding']
                     encode_base64(new)
             except MessageAttachment.DoesNotExist:
@@ -712,7 +788,7 @@ class Message(models.Model):
         """Returns an `email.message.EmailMessage` instance representing the
         contents of this message and all attachments.
 
-        See [email.message.EmailMessage]_ for more information as to what methods
+        See [email.message.EmailMessage]_ for more information like what methods
         and properties are available on `email.message.EmailMessage` instances.
 
         .. note::
@@ -732,9 +808,8 @@ class Message(models.Model):
                 if self.eml.name.endswith('.gz'):
                     body = gzip.GzipFile(fileobj=self.eml).read()
                 else:
-                    self.eml.open()
-                    body = self.eml.file.read()
-                    self.eml.close()
+                    with self.eml.open():
+                        body = self.eml.file.read()
             else:
                 body = self.get_body()
             flat = email.message_from_bytes(body)
@@ -789,6 +864,11 @@ class MessageAttachment(models.Model):
         return email.message_from_string(headers)
 
     def _set_dehydrated_headers(self, email_object):
+        if email_object._payload is None:
+            # otherwise it breaks in
+            # lib/python3.x/email/generator.py:_handle_message,
+            # line: "self._fp.write(payload)"
+            email_object._payload = ""
         self.headers = email_object.as_string()
 
     def __delitem__(self, name):
@@ -822,7 +902,9 @@ class MessageAttachment(models.Model):
         return value
 
     def __str__(self):
-        return self.document.url
+        if self.document:
+            return f'{self.get_filename()}: {self.document.url}'
+        return self.get_filename()
 
     class Meta:
         verbose_name = _('Message attachment')
